+++ 2+2W.litmus
@@ -1,0 +1,14 @@
+X86 2+2W
+"PodWW Wse PodWW Wse"
+Cycle=Wse PodWW Wse PodWW
+Relax=
+Safe=Wse PodWW
+Prefetch=0:x=F,0:y=W,1:y=F,1:x=W
+Com=Ws Ws
+Orig=PodWW Wse PodWW Wse
+{
+}
+ P0         | P1         ;
+ MOV [x],$2 | MOV [y],$2 ;
+ MOV [y],$1 | MOV [x],$1 ;
+forall x=y+1 /\ (y=2 \/ y=1) \/ x=2 /\ y=1
