+++ dartagnan/src/main/java/com/dat3m/dartagnan/exception/ProgramProcessingException.java
@@ -1,0 +1,8 @@
+package com.dat3m.dartagnan.exception;
+
+public class ProgramProcessingException extends IllegalStateException {
+
+    public ProgramProcessingException(String msg){
+        super(msg);
+    }
+}
