--- dartagnan/src/main/java/com/dat3m/dartagnan/parsers/cat/utils/ParsingException.java
@@ -1,8 +1,0 @@
-package com.dat3m.dartagnan.parsers.cat.utils;
-
-public class ParsingException extends RuntimeException {
-
-    public ParsingException(String ctx){
-        super("Invalid syntax at " + ctx);
-    }
-}
