--- dartagnan/src/main/java/com/dat3m/dartagnan/parsers/program/exception/AlreadyExistentVariableException.java
@@ -1,8 +1,0 @@
-package com.dat3m.dartagnan.parsers.program.exception;
-
-public class AlreadyExistentVariableException extends MalformedProgramException {
-
-    public AlreadyExistentVariableException(String msg){
-        super(msg);
-    }
-}
