--- dartagnan/src/main/java/com/dat3m/dartagnan/parsers/program/exception/ExprTypeMismatchException.java
@@ -1,8 +1,0 @@
-package com.dat3m.dartagnan.parsers.program.exception;
-
-public class ExprTypeMismatchException extends RuntimeException {
-
-    public ExprTypeMismatchException(String msg){
-        super(msg);
-    }
-}
