--- dartagnan/src/main/java/com/dat3m/dartagnan/parsers/program/exception/UninitialisedVariableException.java
@@ -1,8 +1,0 @@
-package com.dat3m.dartagnan.parsers.program.exception;
-
-public class UninitialisedVariableException extends MalformedProgramException {
-
-    public UninitialisedVariableException(String msg){
-        super(msg);
-    }
-}
