--- dartagnan/src/main/java/com/dat3m/dartagnan/parsers/program/exception/UnrollException.java
@@ -1,8 +1,0 @@
-package com.dat3m.dartagnan.parsers.program.exception;
-
-public class UnrollException extends IllegalStateException {
-
-    public UnrollException(String msg){
-        super(msg);
-    }
-}
