+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/arch/linux/event/cond/FenceCond.java
@@ -1,6 +1,6 @@
 package com.dat3m.dartagnan.program.arch.linux.event.cond;
 
-import com.dat3m.dartagnan.parsers.program.exception.UnrollException;
+import com.dat3m.dartagnan.exception.ProgramProcessingException;
 import com.dat3m.dartagnan.program.event.Event;
 import com.dat3m.dartagnan.program.event.Fence;
 import com.dat3m.dartagnan.utils.recursion.RecursiveAction;
@@ -49,6 +49,6 @@ public class FenceCond extends Fence {
 
     @Override
     protected RecursiveAction unrollRecursive(int bound, Event predecessor, int depth) {
-        throw new UnrollException("FenceCond cannot be unrolled: event must be generated during compilation");
+        throw new ProgramProcessingException("FenceCond cannot be unrolled: event must be generated during compilation");
     }
 }
