+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/ExecutionStatus.java
@@ -1,7 +1,7 @@
 package com.dat3m.dartagnan.program.event;
 
 import com.dat3m.dartagnan.expression.IConst;
-import com.dat3m.dartagnan.parsers.program.exception.UnrollException;
+import com.dat3m.dartagnan.exception.ProgramProcessingException;
 import com.dat3m.dartagnan.program.Register;
 import com.dat3m.dartagnan.program.event.utils.RegWriter;
 import com.dat3m.dartagnan.program.utils.EType;
@@ -68,6 +68,6 @@ public class ExecutionStatus extends Event implements RegWriter {
 
     @Override
     protected RecursiveAction unrollRecursive(int bound, Event predecessor, int depth) {
-        throw new UnrollException("ExecutionStatus cannot be unrolled: event must be generated during compilation");
+        throw new ProgramProcessingException("ExecutionStatus cannot be unrolled: event must be generated during compilation");
     }
 }
