+++ dartagnan/src/test/java/com/dat3m/dartagnan/parsers/program/utils/ProgramBuilderTest.java
@@ -4,7 +4,7 @@ import com.dat3m.dartagnan.parsers.program.ProgramParser;
 import com.dat3m.dartagnan.utils.ResourceHelper;
 import org.junit.Test;
 
-import com.dat3m.dartagnan.parsers.program.exception.MalformedProgramException;
+import com.dat3m.dartagnan.exception.MalformedProgramException;
 
 import java.io.File;
 import java.io.IOException;
